requirements.txt
Flask==2.2.4
grpcio==1.64.1
python-decouple==3.8
protobuf==4.22.3
